Moved QR generation onto the pages using it
