+++ QRCodeAuth-Web/QRCodeAuth-Web/RequestCredentials.aspx.cs
@@ -7,6 +7,7 @@ using QRCodeAuth_Web.Models;
 using System.Collections.Generic;
 using Newtonsoft.Json;
 using System.Web.UI.WebControls;
+using ZXing;
 
 namespace QRCodeAuth_Web
 {
@@ -21,19 +22,20 @@ namespace QRCodeAuth_Web
 		{
             //TESTING - DELETE LATER
             giveValueToAccounts();
+			imageQRCode.Visible = false;
 
-            //GetLoggedInUserInfo();
-        }
+			//GetLoggedInUserInfo();
+		}
 
         //If confirm butoon is clicked direct user to the QRCode page to display their QRCode. 
         protected void btnConfirm_Click(object sender, EventArgs e)
         {
-            //string shareCredentialObject = getShareCredentialObject();
-            Session["QRString"] = getShareCredentialObject(); 
+			string shareCredentialObject = getShareCredentialObject();
+			generateQRCode(shareCredentialObject);
 
-           Response.Redirect("QRCodePage.aspx");
+			btnCancel.Text = "Done";
+			btnConfirm.Visible = false;
         }
-
 
         //If cancel button is clicked redirect user back to home page. 
         protected void btnCancel_Click(object sender, EventArgs e)
@@ -41,9 +43,56 @@ namespace QRCodeAuth_Web
             Response.Redirect("Home.aspx");
         }
 
+		//Will return a list of the requested credential types selected by information collector. 
+		public List<CredentialType> getRequestedCredentialTypes()
+		{
+			List<CredentialType> types = new List<CredentialType>();
 
-        //Wil return the object that will be encoded into QRCode. 
-        public string getShareCredentialObject()
+			foreach (ListItem item in cblRequestedCredentials.Items)
+			{
+				if (item.Selected)
+				{
+					string selection = item.Value;
+					System.Diagnostics.Debug.WriteLine(selection);
+
+					switch (selection)
+					{
+						case "Name":
+							types.Add(CredentialType.Name);
+							break;
+						case "Email":
+							types.Add(CredentialType.Email);
+							break;
+						case "ID Number":
+							types.Add(CredentialType.IdNumber);
+							break;
+						case "Date of Birth":
+							types.Add(CredentialType.Birthdate);
+							break;
+						case "Address":
+							types.Add(CredentialType.Address);
+							break;
+						case "PhoneNumber":
+							types.Add(CredentialType.PhoneNumber);
+							break;
+						case "Major":
+							types.Add(CredentialType.Major);
+							break;
+						case "Classification":
+							types.Add(CredentialType.Classification);
+							break;
+						case "Work Title":
+							types.Add(CredentialType.WorkTitle);
+							break;
+					}
+				}
+			}
+			cblRequestedCredentials.Visible = false;
+			return types;
+		}
+
+		//Wil return the object that will be encoded into QRCode. 
+		public string getShareCredentialObject()
         {
 			// Anonymous object
             var credentialRequest = new
@@ -58,56 +107,22 @@ namespace QRCodeAuth_Web
             return result;
         }
 
+		protected void generateQRCode(string qrCodeString)
+		{
+			//get currentPath
+			string path = AppDomain.CurrentDomain.BaseDirectory;
 
-        //Will return a list of the requested credential types selected by information collector. 
-        public List<CredentialType> getRequestedCredentialTypes()
-        {
-            List<CredentialType> types = new List<CredentialType>();
+			//Create barcode writer 
+			BarcodeWriter writer = new BarcodeWriter();
+			writer.Format = BarcodeFormat.QR_CODE;
 
-            foreach (ListItem item in cblRequestedCredentials.Items)
-            {
-                if(item.Selected)
-                {
-                    string selection = item.Value;
-                    System.Diagnostics.Debug.WriteLine(selection);
+			//Save the QRCode 
+			writer.Write(qrCodeString).Save(path + @"Images\QRCodes\credentialQR.jpg");
 
-                    switch (selection)
-                    {
-                        case "Name":
-                            types.Add(CredentialType.Name);
-                            break;
-                        case "Email":
-                            types.Add(CredentialType.Email);
-                            break;
-                        case "ID Number":
-                            types.Add(CredentialType.IdNumber);
-                            break;
-                        case "Date of Birth":
-                            types.Add(CredentialType.Birthdate);
-                            break;
-                        case "Address":
-                            types.Add(CredentialType.Address);
-                            break;
-                        case "PhoneNumber":
-                            types.Add(CredentialType.PhoneNumber);
-                            break;
-                        case "Major":
-                            types.Add(CredentialType.Major);
-                            break;
-                        case "Classification":
-                            types.Add(CredentialType.Classification);
-                            break;
-                        case "Work Title":
-                            types.Add(CredentialType.WorkTitle);
-                            break;
-                    }
-                }
-
-            }
-			cblRequestedCredentials.Visible = false;
-			return types;		
-        }
-
+			//Dispaly QRCode
+			imageQRCode.ImageUrl = path + @"Images\QRCodes\credentialQR.jpg";
+			imageQRCode.Visible = true;
+		}
 
         //TESTING - DELETE LATER AND USE SESSION USER AND WEB LOG IN INFO.
         public void giveValueToAccounts()
